+++ 2001/ze_samosbor/samosbor.js
@@ -1,454 +1,448 @@
 import { Entity, Instance } from "cs_script/point_script";
 
 class MathUtils {
-    static clamp(value, min, max) {
-        return Math.min(Math.max(value, min), max);
-    }
+	static clamp(value, min, max) {
+		return Math.min(Math.max(value, min), max);
+	}
 }
 
 const RAD_TO_DEG = 180 / Math.PI;
 
 class Vector3Utils {
-    static equals(a, b) {
-        return a.x === b.x && a.y === b.y && a.z === b.z;
-    }
-    static add(a, b) {
-        return new Vec3(a.x + b.x, a.y + b.y, a.z + b.z);
-    }
-    static subtract(a, b) {
-        return new Vec3(a.x - b.x, a.y - b.y, a.z - b.z);
-    }
-    static scale(vector, scale) {
-        return new Vec3(vector.x * scale, vector.y * scale, vector.z * scale);
-    }
-    static multiply(a, b) {
-        return new Vec3(a.x * b.x, a.y * b.y, a.z * b.z);
-    }
-    static divide(vector, divider) {
-        if (typeof divider === 'number') {
-            if (divider === 0)
-                throw Error('Division by zero');
-            return new Vec3(vector.x / divider, vector.y / divider, vector.z / divider);
-        }
-        else {
-            if (divider.x === 0 || divider.y === 0 || divider.z === 0)
-                throw Error('Division by zero');
-            return new Vec3(vector.x / divider.x, vector.y / divider.y, vector.z / divider.z);
-        }
-    }
-    static length(vector) {
-        return Math.sqrt(Vector3Utils.lengthSquared(vector));
-    }
-    static lengthSquared(vector) {
-        return vector.x ** 2 + vector.y ** 2 + vector.z ** 2;
-    }
-    static length2D(vector) {
-        return Math.sqrt(Vector3Utils.length2DSquared(vector));
-    }
-    static length2DSquared(vector) {
-        return vector.x ** 2 + vector.y ** 2;
-    }
-    static normalize(vector) {
-        const length = Vector3Utils.length(vector);
-        return length ? Vector3Utils.divide(vector, length) : Vec3.Zero;
-    }
-    static dot(a, b) {
-        return a.x * b.x + a.y * b.y + a.z * b.z;
-    }
-    static cross(a, b) {
-        return new Vec3(a.y * b.z - a.z * b.y, a.z * b.x - a.x * b.z, a.x * b.y - a.y * b.x);
-    }
-    static inverse(vector) {
-        return new Vec3(-vector.x, -vector.y, -vector.z);
-    }
-    static distance(a, b) {
-        return Vector3Utils.subtract(a, b).length;
-    }
-    static distanceSquared(a, b) {
-        return Vector3Utils.subtract(a, b).lengthSquared;
-    }
-    static floor(vector) {
-        return new Vec3(Math.floor(vector.x), Math.floor(vector.y), Math.floor(vector.z));
-    }
-    static vectorAngles(vector) {
-        let yaw = 0;
-        let pitch = 0;
-        if (!vector.y && !vector.x) {
-            if (vector.z > 0)
-                pitch = -90;
-            else
-                pitch = 90;
-        }
-        else {
-            yaw = Math.atan2(vector.y, vector.x) * RAD_TO_DEG;
-            pitch = Math.atan2(-vector.z, Vector3Utils.length2D(vector)) * RAD_TO_DEG;
-        }
-        return new Euler({
-            pitch,
-            yaw,
-            roll: 0,
-        });
-    }
-    static lerp(a, b, fraction, clamp = true) {
-        let t = fraction;
-        if (clamp) {
-            t = MathUtils.clamp(t, 0, 1);
-        }
-        // a + (b - a) * t
-        return new Vec3(a.x + (b.x - a.x) * t, a.y + (b.y - a.y) * t, a.z + (b.z - a.z) * t);
-    }
-    static directionTowards(a, b) {
-        return Vector3Utils.subtract(b, a).normal;
-    }
-    static lookAt(a, b) {
-        return Vector3Utils.directionTowards(a, b).eulerAngles;
-    }
-    static withX(vector, x) {
-        return new Vec3(x, vector.y, vector.z);
-    }
-    static withY(vector, y) {
-        return new Vec3(vector.x, y, vector.z);
-    }
-    static withZ(vector, z) {
-        return new Vec3(vector.x, vector.y, z);
-    }
+	static equals(a, b) {
+		return a.x === b.x && a.y === b.y && a.z === b.z;
+	}
+	static add(a, b) {
+		return new Vec3(a.x + b.x, a.y + b.y, a.z + b.z);
+	}
+	static subtract(a, b) {
+		return new Vec3(a.x - b.x, a.y - b.y, a.z - b.z);
+	}
+	static scale(vector, scale) {
+		return new Vec3(vector.x * scale, vector.y * scale, vector.z * scale);
+	}
+	static multiply(a, b) {
+		return new Vec3(a.x * b.x, a.y * b.y, a.z * b.z);
+	}
+	static divide(vector, divider) {
+		if (typeof divider === 'number') {
+			if (divider === 0)
+				throw Error('Division by zero');
+			return new Vec3(vector.x / divider, vector.y / divider, vector.z / divider);
+		}
+		else {
+			if (divider.x === 0 || divider.y === 0 || divider.z === 0)
+				throw Error('Division by zero');
+			return new Vec3(vector.x / divider.x, vector.y / divider.y, vector.z / divider.z);
+		}
+	}
+	static length(vector) {
+		return Math.sqrt(Vector3Utils.lengthSquared(vector));
+	}
+	static lengthSquared(vector) {
+		return vector.x ** 2 + vector.y ** 2 + vector.z ** 2;
+	}
+	static length2D(vector) {
+		return Math.sqrt(Vector3Utils.length2DSquared(vector));
+	}
+	static length2DSquared(vector) {
+		return vector.x ** 2 + vector.y ** 2;
+	}
+	static normalize(vector) {
+		const length = Vector3Utils.length(vector);
+		return length ? Vector3Utils.divide(vector, length) : Vec3.Zero;
+	}
+	static dot(a, b) {
+		return a.x * b.x + a.y * b.y + a.z * b.z;
+	}
+	static cross(a, b) {
+		return new Vec3(a.y * b.z - a.z * b.y, a.z * b.x - a.x * b.z, a.x * b.y - a.y * b.x);
+	}
+	static inverse(vector) {
+		return new Vec3(-vector.x, -vector.y, -vector.z);
+	}
+	static distance(a, b) {
+		return Vector3Utils.subtract(a, b).length;
+	}
+	static distanceSquared(a, b) {
+		return Vector3Utils.subtract(a, b).lengthSquared;
+	}
+	static floor(vector) {
+		return new Vec3(Math.floor(vector.x), Math.floor(vector.y), Math.floor(vector.z));
+	}
+	static vectorAngles(vector) {
+		let yaw = 0;
+		let pitch = 0;
+		if (!vector.y && !vector.x) {
+			if (vector.z > 0)
+				pitch = -90;
+			else
+				pitch = 90;
+		}
+		else {
+			yaw = Math.atan2(vector.y, vector.x) * RAD_TO_DEG;
+			pitch = Math.atan2(-vector.z, Vector3Utils.length2D(vector)) * RAD_TO_DEG;
+		}
+		return new Euler({
+			pitch,
+			yaw,
+			roll: 0,
+		});
+	}
+	static lerp(a, b, fraction, clamp = true) {
+		let t = fraction;
+		if (clamp) {
+			t = MathUtils.clamp(t, 0, 1);
+		}
+		// a + (b - a) * t
+		return new Vec3(a.x + (b.x - a.x) * t, a.y + (b.y - a.y) * t, a.z + (b.z - a.z) * t);
+	}
+	static directionTowards(a, b) {
+		return Vector3Utils.subtract(b, a).normal;
+	}
+	static lookAt(a, b) {
+		return Vector3Utils.directionTowards(a, b).eulerAngles;
+	}
+	static withX(vector, x) {
+		return new Vec3(x, vector.y, vector.z);
+	}
+	static withY(vector, y) {
+		return new Vec3(vector.x, y, vector.z);
+	}
+	static withZ(vector, z) {
+		return new Vec3(vector.x, vector.y, z);
+	}
 }
 class Vec3 {
-    x;
-    y;
-    z;
-    static Zero = new Vec3(0, 0, 0);
-    constructor(xOrVector, y, z) {
-        if (typeof xOrVector === 'object') {
-            this.x = xOrVector.x === 0 ? 0 : xOrVector.x;
-            this.y = xOrVector.y === 0 ? 0 : xOrVector.y;
-            this.z = xOrVector.z === 0 ? 0 : xOrVector.z;
-        }
-        else {
-            this.x = xOrVector === 0 ? 0 : xOrVector;
-            this.y = y === 0 ? 0 : y;
-            this.z = z === 0 ? 0 : z;
-        }
-    }
-    get length() {
-        return Vector3Utils.length(this);
-    }
-    get lengthSquared() {
-        return Vector3Utils.lengthSquared(this);
-    }
-    get length2D() {
-        return Vector3Utils.length2D(this);
-    }
-    get length2DSquared() {
-        return Vector3Utils.length2DSquared(this);
-    }
-    /**
-     * Normalizes the vector (Dividing the vector by its length to have the length be equal to 1 e.g. [0.0, 0.666, 0.333])
-     */
-    get normal() {
-        return Vector3Utils.normalize(this);
-    }
-    get inverse() {
-        return Vector3Utils.inverse(this);
-    }
-    /**
-     * Floor (Round down) each vector component
-     */
-    get floored() {
-        return Vector3Utils.floor(this);
-    }
-    /**
-     * Calculates the angles from a forward vector
-     */
-    get eulerAngles() {
-        return Vector3Utils.vectorAngles(this);
-    }
-    toString() {
-        return `Vec3: [${this.x}, ${this.y}, ${this.z}]`;
-    }
-    equals(vector) {
-        return Vector3Utils.equals(this, vector);
-    }
-    add(vector) {
-        return Vector3Utils.add(this, vector);
-    }
-    subtract(vector) {
-        return Vector3Utils.subtract(this, vector);
-    }
-    divide(vector) {
-        return Vector3Utils.divide(this, vector);
-    }
-    scale(scaleOrVector) {
-        return typeof scaleOrVector === 'number'
-            ? Vector3Utils.scale(this, scaleOrVector)
-            : Vector3Utils.multiply(this, scaleOrVector);
-    }
-    multiply(scaleOrVector) {
-        return typeof scaleOrVector === 'number'
-            ? Vector3Utils.scale(this, scaleOrVector)
-            : Vector3Utils.multiply(this, scaleOrVector);
-    }
-    dot(vector) {
-        return Vector3Utils.dot(this, vector);
-    }
-    cross(vector) {
-        return Vector3Utils.cross(this, vector);
-    }
-    distance(vector) {
-        return Vector3Utils.distance(this, vector);
-    }
-    distanceSquared(vector) {
-        return Vector3Utils.distanceSquared(this, vector);
-    }
-    /**
-     * Linearly interpolates the vector to a point based on a 0.0-1.0 fraction
-     * Clamp limits the fraction to [0,1]
-     */
-    lerpTo(vector, fraction, clamp = true) {
-        return Vector3Utils.lerp(this, vector, fraction, clamp);
-    }
-    /**
-     * Gets the normalized direction vector pointing towards specified point (subtracting two vectors)
-     */
-    directionTowards(vector) {
-        return Vector3Utils.directionTowards(this, vector);
-    }
-    /**
-     * Returns an angle pointing towards a point from the current vector
-     */
-    lookAt(vector) {
-        return Vector3Utils.lookAt(this, vector);
-    }
-    /**
-     * Returns the same vector but with a supplied X component
-     */
-    withX(x) {
-        return Vector3Utils.withX(this, x);
-    }
-    /**
-     * Returns the same vector but with a supplied Y component
-     */
-    withY(y) {
-        return Vector3Utils.withY(this, y);
-    }
-    /**
-     * Returns the same vector but with a supplied Z component
-     */
-    withZ(z) {
-        return Vector3Utils.withZ(this, z);
-    }
+	x;
+	y;
+	z;
+	static Zero = new Vec3(0, 0, 0);
+	constructor(xOrVector, y, z) {
+		if (typeof xOrVector === 'object') {
+			this.x = xOrVector.x === 0 ? 0 : xOrVector.x;
+			this.y = xOrVector.y === 0 ? 0 : xOrVector.y;
+			this.z = xOrVector.z === 0 ? 0 : xOrVector.z;
+		}
+		else {
+			this.x = xOrVector === 0 ? 0 : xOrVector;
+			this.y = y === 0 ? 0 : y;
+			this.z = z === 0 ? 0 : z;
+		}
+	}
+	get length() {
+		return Vector3Utils.length(this);
+	}
+	get lengthSquared() {
+		return Vector3Utils.lengthSquared(this);
+	}
+	get length2D() {
+		return Vector3Utils.length2D(this);
+	}
+	get length2DSquared() {
+		return Vector3Utils.length2DSquared(this);
+	}
+	/**
+	 * Normalizes the vector (Dividing the vector by its length to have the length be equal to 1 e.g. [0.0, 0.666, 0.333])
+	 */
+	get normal() {
+		return Vector3Utils.normalize(this);
+	}
+	get inverse() {
+		return Vector3Utils.inverse(this);
+	}
+	/**
+	 * Floor (Round down) each vector component
+	 */
+	get floored() {
+		return Vector3Utils.floor(this);
+	}
+	/**
+	 * Calculates the angles from a forward vector
+	 */
+	get eulerAngles() {
+		return Vector3Utils.vectorAngles(this);
+	}
+	toString() {
+		return `Vec3: [${this.x}, ${this.y}, ${this.z}]`;
+	}
+	equals(vector) {
+		return Vector3Utils.equals(this, vector);
+	}
+	add(vector) {
+		return Vector3Utils.add(this, vector);
+	}
+	subtract(vector) {
+		return Vector3Utils.subtract(this, vector);
+	}
+	divide(vector) {
+		return Vector3Utils.divide(this, vector);
+	}
+	scale(scaleOrVector) {
+		return typeof scaleOrVector === 'number'
+			? Vector3Utils.scale(this, scaleOrVector)
+			: Vector3Utils.multiply(this, scaleOrVector);
+	}
+	multiply(scaleOrVector) {
+		return typeof scaleOrVector === 'number'
+			? Vector3Utils.scale(this, scaleOrVector)
+			: Vector3Utils.multiply(this, scaleOrVector);
+	}
+	dot(vector) {
+		return Vector3Utils.dot(this, vector);
+	}
+	cross(vector) {
+		return Vector3Utils.cross(this, vector);
+	}
+	distance(vector) {
+		return Vector3Utils.distance(this, vector);
+	}
+	distanceSquared(vector) {
+		return Vector3Utils.distanceSquared(this, vector);
+	}
+	/**
+	 * Linearly interpolates the vector to a point based on a 0.0-1.0 fraction
+	 * Clamp limits the fraction to [0,1]
+	 */
+	lerpTo(vector, fraction, clamp = true) {
+		return Vector3Utils.lerp(this, vector, fraction, clamp);
+	}
+	/**
+	 * Gets the normalized direction vector pointing towards specified point (subtracting two vectors)
+	 */
+	directionTowards(vector) {
+		return Vector3Utils.directionTowards(this, vector);
+	}
+	/**
+	 * Returns an angle pointing towards a point from the current vector
+	 */
+	lookAt(vector) {
+		return Vector3Utils.lookAt(this, vector);
+	}
+	/**
+	 * Returns the same vector but with a supplied X component
+	 */
+	withX(x) {
+		return Vector3Utils.withX(this, x);
+	}
+	/**
+	 * Returns the same vector but with a supplied Y component
+	 */
+	withY(y) {
+		return Vector3Utils.withY(this, y);
+	}
+	/**
+	 * Returns the same vector but with a supplied Z component
+	 */
+	withZ(z) {
+		return Vector3Utils.withZ(this, z);
+	}
 }
 
 class EulerUtils {
-    static equals(a, b) {
-        return a.pitch === b.pitch && a.yaw === b.yaw && a.roll === b.roll;
-    }
-    static normalize(angle) {
-        const normalizeAngle = (angle) => {
-            angle = angle % 360;
-            if (angle > 180)
-                return angle - 360;
-            if (angle < -180)
-                return angle + 360;
-            return angle;
-        };
-        return new Euler(normalizeAngle(angle.pitch), normalizeAngle(angle.yaw), normalizeAngle(angle.roll));
-    }
-    static forward(angle) {
-        const pitchInRad = (angle.pitch / 180) * Math.PI;
-        const yawInRad = (angle.yaw / 180) * Math.PI;
-        const cosPitch = Math.cos(pitchInRad);
-        return new Vec3(cosPitch * Math.cos(yawInRad), cosPitch * Math.sin(yawInRad), -Math.sin(pitchInRad));
-    }
-    static right(angle) {
-        const pitchInRad = (angle.pitch / 180) * Math.PI;
-        const yawInRad = (angle.yaw / 180) * Math.PI;
-        const rollInRad = (angle.roll / 180) * Math.PI;
-        const sinPitch = Math.sin(pitchInRad);
-        const sinYaw = Math.sin(yawInRad);
-        const sinRoll = Math.sin(rollInRad);
-        const cosPitch = Math.cos(pitchInRad);
-        const cosYaw = Math.cos(yawInRad);
-        const cosRoll = Math.cos(rollInRad);
-        return new Vec3(-1 * sinRoll * sinPitch * cosYaw + -1 * cosRoll * -sinYaw, -1 * sinRoll * sinPitch * sinYaw + -1 * cosRoll * cosYaw, -1 * sinRoll * cosPitch);
-    }
-    static up(angle) {
-        const pitchInRad = (angle.pitch / 180) * Math.PI;
-        const yawInRad = (angle.yaw / 180) * Math.PI;
-        const rollInRad = (angle.roll / 180) * Math.PI;
-        const sinPitch = Math.sin(pitchInRad);
-        const sinYaw = Math.sin(yawInRad);
-        const sinRoll = Math.sin(rollInRad);
-        const cosPitch = Math.cos(pitchInRad);
-        const cosYaw = Math.cos(yawInRad);
-        const cosRoll = Math.cos(rollInRad);
-        return new Vec3(cosRoll * sinPitch * cosYaw + -sinRoll * -sinYaw, cosRoll * sinPitch * sinYaw + -sinRoll * cosYaw, cosRoll * cosPitch);
-    }
-    static lerp(a, b, fraction, clamp = true) {
-        let t = fraction;
-        if (clamp) {
-            t = MathUtils.clamp(t, 0, 1);
-        }
-        const lerpComponent = (start, end, t) => {
-            // Calculate the shortest angular distance
-            let delta = end - start;
-            // Normalize delta to [-180, 180] range to find shortest path
-            if (delta > 180) {
-                delta -= 360;
-            }
-            else if (delta < -180) {
-                delta += 360;
-            }
-            // Interpolate using the shortest path
-            return start + delta * t;
-        };
-        // a + (b - a) * t
-        return new Euler(lerpComponent(a.pitch, b.pitch, t), lerpComponent(a.yaw, b.yaw, t), lerpComponent(a.roll, b.roll, t));
-    }
-    static withPitch(angle, pitch) {
-        return new Euler(pitch, angle.yaw, angle.roll);
-    }
-    static withYaw(angle, yaw) {
-        return new Euler(angle.pitch, yaw, angle.roll);
-    }
-    static withRoll(angle, roll) {
-        return new Euler(angle.pitch, angle.yaw, roll);
-    }
-    static rotateTowards(current, target, maxStep) {
-        const rotateComponent = (current, target, step) => {
-            let delta = target - current;
-            if (delta > 180) {
-                delta -= 360;
-            }
-            else if (delta < -180) {
-                delta += 360;
-            }
-            if (Math.abs(delta) <= step) {
-                return target;
-            }
-            else {
-                return current + Math.sign(delta) * step;
-            }
-        };
-        return new Euler(rotateComponent(current.pitch, target.pitch, maxStep), rotateComponent(current.yaw, target.yaw, maxStep), rotateComponent(current.roll, target.roll, maxStep));
-    }
-    static clamp(angle, min, max) {
-        return new Euler(MathUtils.clamp(angle.pitch, min.pitch, max.pitch), MathUtils.clamp(angle.yaw, min.yaw, max.yaw), MathUtils.clamp(angle.roll, min.roll, max.roll));
-    }
+	static equals(a, b) {
+		return a.pitch === b.pitch && a.yaw === b.yaw && a.roll === b.roll;
+	}
+	static normalize(angle) {
+		const normalizeAngle = (angle) => {
+			angle = angle % 360;
+			if (angle > 180)
+				return angle - 360;
+			if (angle < -180)
+				return angle + 360;
+			return angle;
+		};
+		return new Euler(normalizeAngle(angle.pitch), normalizeAngle(angle.yaw), normalizeAngle(angle.roll));
+	}
+	static forward(angle) {
+		const pitchInRad = (angle.pitch / 180) * Math.PI;
+		const yawInRad = (angle.yaw / 180) * Math.PI;
+		const cosPitch = Math.cos(pitchInRad);
+		return new Vec3(cosPitch * Math.cos(yawInRad), cosPitch * Math.sin(yawInRad), -Math.sin(pitchInRad));
+	}
+	static right(angle) {
+		const pitchInRad = (angle.pitch / 180) * Math.PI;
+		const yawInRad = (angle.yaw / 180) * Math.PI;
+		const rollInRad = (angle.roll / 180) * Math.PI;
+		const sinPitch = Math.sin(pitchInRad);
+		const sinYaw = Math.sin(yawInRad);
+		const sinRoll = Math.sin(rollInRad);
+		const cosPitch = Math.cos(pitchInRad);
+		const cosYaw = Math.cos(yawInRad);
+		const cosRoll = Math.cos(rollInRad);
+		return new Vec3(-1 * sinRoll * sinPitch * cosYaw + -1 * cosRoll * -sinYaw, -1 * sinRoll * sinPitch * sinYaw + -1 * cosRoll * cosYaw, -1 * sinRoll * cosPitch);
+	}
+	static up(angle) {
+		const pitchInRad = (angle.pitch / 180) * Math.PI;
+		const yawInRad = (angle.yaw / 180) * Math.PI;
+		const rollInRad = (angle.roll / 180) * Math.PI;
+		const sinPitch = Math.sin(pitchInRad);
+		const sinYaw = Math.sin(yawInRad);
+		const sinRoll = Math.sin(rollInRad);
+		const cosPitch = Math.cos(pitchInRad);
+		const cosYaw = Math.cos(yawInRad);
+		const cosRoll = Math.cos(rollInRad);
+		return new Vec3(cosRoll * sinPitch * cosYaw + -sinRoll * -sinYaw, cosRoll * sinPitch * sinYaw + -sinRoll * cosYaw, cosRoll * cosPitch);
+	}
+	static lerp(a, b, fraction, clamp = true) {
+		let t = fraction;
+		if (clamp) {
+			t = MathUtils.clamp(t, 0, 1);
+		}
+		const lerpComponent = (start, end, t) => {
+			// Calculate the shortest angular distance
+			let delta = end - start;
+			// Normalize delta to [-180, 180] range to find shortest path
+			if (delta > 180) {
+				delta -= 360;
+			}
+			else if (delta < -180) {
+				delta += 360;
+			}
+			// Interpolate using the shortest path
+			return start + delta * t;
+		};
+		// a + (b - a) * t
+		return new Euler(lerpComponent(a.pitch, b.pitch, t), lerpComponent(a.yaw, b.yaw, t), lerpComponent(a.roll, b.roll, t));
+	}
+	static withPitch(angle, pitch) {
+		return new Euler(pitch, angle.yaw, angle.roll);
+	}
+	static withYaw(angle, yaw) {
+		return new Euler(angle.pitch, yaw, angle.roll);
+	}
+	static withRoll(angle, roll) {
+		return new Euler(angle.pitch, angle.yaw, roll);
+	}
+	static rotateTowards(current, target, maxStep) {
+		const rotateComponent = (current, target, step) => {
+			let delta = target - current;
+			if (delta > 180) {
+				delta -= 360;
+			}
+			else if (delta < -180) {
+				delta += 360;
+			}
+			if (Math.abs(delta) <= step) {
+				return target;
+			}
+			else {
+				return current + Math.sign(delta) * step;
+			}
+		};
+		return new Euler(rotateComponent(current.pitch, target.pitch, maxStep), rotateComponent(current.yaw, target.yaw, maxStep), rotateComponent(current.roll, target.roll, maxStep));
+	}
+	static clamp(angle, min, max) {
+		return new Euler(MathUtils.clamp(angle.pitch, min.pitch, max.pitch), MathUtils.clamp(angle.yaw, min.yaw, max.yaw), MathUtils.clamp(angle.roll, min.roll, max.roll));
+	}
 }
 class Euler {
-    pitch;
-    yaw;
-    roll;
-    static Zero = new Euler(0, 0, 0);
-    constructor(pitchOrAngle, yaw, roll) {
-        if (typeof pitchOrAngle === 'object') {
-            this.pitch = pitchOrAngle.pitch === 0 ? 0 : pitchOrAngle.pitch;
-            this.yaw = pitchOrAngle.yaw === 0 ? 0 : pitchOrAngle.yaw;
-            this.roll = pitchOrAngle.roll === 0 ? 0 : pitchOrAngle.roll;
-        }
-        else {
-            this.pitch = pitchOrAngle === 0 ? pitchOrAngle : pitchOrAngle;
-            this.yaw = yaw === 0 ? 0 : yaw;
-            this.roll = roll === 0 ? 0 : roll;
-        }
-    }
-    /**
-     * Returns angle with every componented clamped from -180 to 180
-     */
-    get normal() {
-        return EulerUtils.normalize(this);
-    }
-    /**
-     * Returns a normalized forward direction vector
-     */
-    get forward() {
-        return EulerUtils.forward(this);
-    }
-    /**
-     * Returns a normalized backward direction vector
-     */
-    get backward() {
-        return this.forward.inverse;
-    }
-    /**
-     * Returns a normalized right direction vector
-     */
-    get right() {
-        return EulerUtils.right(this);
-    }
-    /**
-     * Returns a normalized left direction vector
-     */
-    get left() {
-        return this.right.inverse;
-    }
-    /**
-     * Returns a normalized up direction vector
-     */
-    get up() {
-        return EulerUtils.up(this);
-    }
-    /**
-     * Returns a normalized down direction vector
-     */
-    get down() {
-        return this.up.inverse;
-    }
-    toString() {
-        return `Euler: [${this.pitch}, ${this.yaw}, ${this.roll}]`;
-    }
-    equals(angle) {
-        return EulerUtils.equals(this, angle);
-    }
-    /**
-     * Linearly interpolates the angle to an angle based on a 0.0-1.0 fraction
-     * Clamp limits the fraction to [0,1]
-     * ! Euler angles are not suited for interpolation, prefer to use quarternions instead
-     */
-    lerp(angle, fraction, clamp = true) {
-        return EulerUtils.lerp(this, angle, fraction, clamp);
-    }
-    /**
-     * Returns the same angle but with a supplied pitch component
-     */
-    withPitch(pitch) {
-        return EulerUtils.withPitch(this, pitch);
-    }
-    /**
-     * Returns the same angle but with a supplied yaw component
-     */
-    withYaw(yaw) {
-        return EulerUtils.withYaw(this, yaw);
-    }
-    /**
-     * Returns the same angle but with a supplied roll component
-     */
-    withRoll(roll) {
-        return EulerUtils.withRoll(this, roll);
-    }
-    /**
-     * Rotates an angle towards another angle by a specific step
-     * ! Euler angles are not suited for interpolation, prefer to use quarternions instead
-     */
-    rotateTowards(angle, maxStep) {
-        return EulerUtils.rotateTowards(this, angle, maxStep);
-    }
-    /**
-     * Clamps each component (pitch, yaw, roll) between the corresponding min and max values
-     */
-    clamp(min, max) {
-        return EulerUtils.clamp(this, min, max);
-    }
+	pitch;
+	yaw;
+	roll;
+	static Zero = new Euler(0, 0, 0);
+	constructor(pitchOrAngle, yaw, roll) {
+		if (typeof pitchOrAngle === 'object') {
+			this.pitch = pitchOrAngle.pitch === 0 ? 0 : pitchOrAngle.pitch;
+			this.yaw = pitchOrAngle.yaw === 0 ? 0 : pitchOrAngle.yaw;
+			this.roll = pitchOrAngle.roll === 0 ? 0 : pitchOrAngle.roll;
+		}
+		else {
+			this.pitch = pitchOrAngle === 0 ? pitchOrAngle : pitchOrAngle;
+			this.yaw = yaw === 0 ? 0 : yaw;
+			this.roll = roll === 0 ? 0 : roll;
+		}
+	}
+	/**
+	 * Returns angle with every componented clamped from -180 to 180
+	 */
+	get normal() {
+		return EulerUtils.normalize(this);
+	}
+	/**
+	 * Returns a normalized forward direction vector
+	 */
+	get forward() {
+		return EulerUtils.forward(this);
+	}
+	/**
+	 * Returns a normalized backward direction vector
+	 */
+	get backward() {
+		return this.forward.inverse;
+	}
+	/**
+	 * Returns a normalized right direction vector
+	 */
+	get right() {
+		return EulerUtils.right(this);
+	}
+	/**
+	 * Returns a normalized left direction vector
+	 */
+	get left() {
+		return this.right.inverse;
+	}
+	/**
+	 * Returns a normalized up direction vector
+	 */
+	get up() {
+		return EulerUtils.up(this);
+	}
+	/**
+	 * Returns a normalized down direction vector
+	 */
+	get down() {
+		return this.up.inverse;
+	}
+	toString() {
+		return `Euler: [${this.pitch}, ${this.yaw}, ${this.roll}]`;
+	}
+	equals(angle) {
+		return EulerUtils.equals(this, angle);
+	}
+	/**
+	 * Linearly interpolates the angle to an angle based on a 0.0-1.0 fraction
+	 * Clamp limits the fraction to [0,1]
+	 * ! Euler angles are not suited for interpolation, prefer to use quarternions instead
+	 */
+	lerp(angle, fraction, clamp = true) {
+		return EulerUtils.lerp(this, angle, fraction, clamp);
+	}
+	/**
+	 * Returns the same angle but with a supplied pitch component
+	 */
+	withPitch(pitch) {
+		return EulerUtils.withPitch(this, pitch);
+	}
+	/**
+	 * Returns the same angle but with a supplied yaw component
+	 */
+	withYaw(yaw) {
+		return EulerUtils.withYaw(this, yaw);
+	}
+	/**
+	 * Returns the same angle but with a supplied roll component
+	 */
+	withRoll(roll) {
+		return EulerUtils.withRoll(this, roll);
+	}
+	/**
+	 * Rotates an angle towards another angle by a specific step
+	 * ! Euler angles are not suited for interpolation, prefer to use quarternions instead
+	 */
+	rotateTowards(angle, maxStep) {
+		return EulerUtils.rotateTowards(this, angle, maxStep);
+	}
+	/**
+	 * Clamps each component (pitch, yaw, roll) between the corresponding min and max values
+	 */
+	clamp(min, max) {
+		return EulerUtils.clamp(this, min, max);
+	}
 }
 
 const Inputs = [
-    ["Map_Boss_TheHeart_Start", "OnTrigger", "Map_Boss_TheHeart_GetDamage,Map_Script,Map_Boss_Health_Hud,[The$Heart]$[2/2],100,1", StartBoss, 3.00],
-    ["Boss_Phase_Case", "OnCase01", "Map_Boss_TheHeart_GetDamage,Map_Script,Map_Boss_Health_Hud,[The$Heart]$[1/2],50,1", StartBoss, 0.10],
-    ["Map_Boss_TheHeart_Start", "OnTrigger", "0,450", AddHealth, 2.90],
-    ["Filter_Team_Human_BossAdd_P2", "OnPass", "1,580", AddHealth, 0.00],
-    ["Map_Boss_TheHeart_Health_Branch", "OnTrue", "-5", ChangeHealthIt, 0.00],
-    ["Map_Boss_Arena_Bottom_SubtractHealth", "OnStartTouch", "70", ChangeHealthIt, 0.00],
     ["Admin_ResetToDefault", "OnPressed", "", ResetVariables, 0.00],
     ["Admin_HP_Sub1", "OnPressed", "1", ChangeHealth, 0.00],
     ["Admin_HP_Sub5", "OnPressed", "5", ChangeHealth, 0.00],
@@ -578,17 +572,17 @@ let BOTTLE_CHANCE = [
 let ITEM_CHANCE = [
     { value: 0, weight: 30 },   // BEER
     { value: 1, weight: 28 },   // BEANS
-    { value: 2, weight: 25 },   // SPANNER
+    { value: 2, weight: 22 },   // SPANNER
     { value: 3, weight: 12 },   // WHIP
-    { value: 4, weight: 5 }     // FLARE GUN
+    { value: 4, weight: 8 }     // FLARE GUN
 ]
 
 let GIFTBOX_CHANCE = [
     { value: 0, weight: 25 },   // BEER
     { value: 1, weight: 10 },   // BEANS
-    { value: 2, weight: 55 },   // MINE
+    { value: 2, weight: 52 },   // MINE
     { value: 3, weight: 5 },    // WHIP
-    { value: 4, weight: 5 }     // FLARE GUN
+    { value: 4, weight: 8 }     // FLARE GUN
 ]
 
 const DelayedCalls = [];
@@ -927,7 +921,6 @@ Instance.OnPlayerReset((event) => {
 
 Instance.OnRoundStart(() => {
     //Instance.EntFireAtName({ name: SCRIPT_ENT, input: "RunScriptInput", value: "GetData" })
-    ResetBossS();
     //UpdateMapStats();
     ResetScript();
     DelayedCalls.length = 0;
@@ -963,7 +956,6 @@ Instance.OnRoundEnd(() => {
     //Instance.EntFireAtName({ name: SCRIPT_ENT, input: "RunScriptInput", value: "SaveData" })
     DelayedCalls.length = 0;
     ResetScript();
-    ResetBossS();
 });
 
 Instance.OnModifyPlayerDamage((event) => {
@@ -2822,15 +2814,22 @@ Instance.OnScriptInput("WormDie", () => {
     MINI_BOSS = "";
     isMusicPick = true;
 
+    Instance.EntFireAtName({ name: "Worm_Face_MineMaker", input: "ClearParent", delay: 0.00 })
+    let meat_random = GetRandomNumber(1, 7)
+    if(meat_random > 1)
+    {
+        Instance.EntFireAtName({ name: "Worm_Face_MineMaker", input: "KeyValue", value: "EntityTemplate Item_Meat_Template", delay: 0.02 })
+        Instance.EntFireAtName({ name: "Worm_Face_MineMaker", input: "ForceSpawn", delay: 0.05 })
+    }
     Instance.EntFireAtName({ name: "Worm_Face_Sound*", input: "FireUser1", delay: 0.00 })
     Instance.EntFireAtName({ name: "Worm_Face_Sound*", input: "StopSound", delay: 0.02 })
     Instance.EntFireAtName({ name: "Worm_Face_Sound*", input: "Kill", delay: 0.04 })
-    Instance.EntFireAtName({ name: "worm_train_*", input: "Kill", delay: 0.05 })
-    Instance.EntFireAtName({ name: "worm_face_*", input: "Kill", delay: 0.05 })
-    Instance.EntFireAtName({ name: "worm_top_*", input: "Kill", delay: 0.05 })
-    Instance.EntFireAtName({ name: "worm_middle_*", input: "Kill", delay: 0.05 })
-    Instance.EntFireAtName({ name: "worm_small_*", input: "Kill", delay: 0.05 })
-    Instance.EntFireAtName({ name: "worm_end_*", input: "Kill", delay: 0.05 })
+    Instance.EntFireAtName({ name: "worm_train_*", input: "Kill", delay: 0.10 })
+    Instance.EntFireAtName({ name: "worm_face_*", input: "Kill", delay: 0.10 })
+    Instance.EntFireAtName({ name: "worm_top_*", input: "Kill", delay: 0.10 })
+    Instance.EntFireAtName({ name: "worm_middle_*", input: "Kill", delay: 0.10 })
+    Instance.EntFireAtName({ name: "worm_small_*", input: "Kill", delay: 0.10 })
+    Instance.EntFireAtName({ name: "worm_end_*", input: "Kill", delay: 0.10 })
     Instance.EntFireAtName({ name: "Worm_Face_ShakeTimer*", input: "Kill", delay: 0.05 })
     Instance.EntFireAtName({ name: "Item_Mine_Button*", input: "Kill", delay: 0.05 })
     Instance.EntFireAtName({ name: "Item_Mine_Maker*", input: "Kill", delay: 0.05 })
@@ -2933,272 +2932,6 @@ Instance.OnScriptInput("WormTick", () => {
     }
 })
 
-//    ___                     _____ _                                  _   
-//   / __\ ___  ___ ___   _  /__   \ |__   ___    /\  /\___  __ _ _ __| |_ 
-//  /__\/// _ \/ __/ __| (_)   / /\/ '_ \ / _ \  / /_/ / _ \/ _` | '__| __|
-// / \/  \ (_) \__ \__ \  _   / /  | | | |  __/ / __  /  __/ (_| | |  | |_ 
-// \_____/\___/|___/___/ (_)  \/   |_| |_|\___| \/ /_/ \___|\__,_|_|   \__|
-
-let BOSS_HEALTH = 0.00;
-let BOSS_MAX_HEALTH = 0.00;
-let HP_BAR_MAX_FRAME = 15;
-let HP_BAR_FRAME = 0;
-let HP_PER_FRAME = 0;
-
-let BOSS_NAME = "BOSS: ";
-let BOSS_ENT = "";
-let BOSS_SCRIPT = "";
-let BOSS_HUD_ENT = "";
-let BOSS_HUD_TEXT = "";
-let BOSS_PERCENT_C = ""; 
-
-let BOSS_HUD_IND = true; 
-let BOSS_HUD_ST = "◼";
-let BOSS_HUD_ST2 = "◻";
-
-let TICKRATE_B = 0.01;
-let IS_BOSS_FIGHT = false;
-
-let ITEM_DAMAGE = "";
-let ITEM_DAMAGE_TICK = 2.00;
-let SAVE_ITEM_DAMAG_T = ITEM_DAMAGE_TICK;
-
-let GRENADE_DAMAGE = 0;
-let GRENADE_DAMAGE_TICK = 2.00;
-let SAVE_GRENADE_DAMAG_T = GRENADE_DAMAGE_TICK;
-
-function StartBoss(arg) 
-{
-    ITEM_DAMAGE = "";
-    GRENADE_DAMAGE = 0;
-    let arg_s = arg;
-    let arg_rs = arg_s.replace(/\s+/g, '');
-    const arr = arg_rs.split(",");
-    BOSS_ENT = arr[0];
-    BOSS_SCRIPT = arr[1];
-    BOSS_HUD_ENT = arr[2];
-    BOSS_NAME = arr[3];
-    if (BOSS_NAME.includes('$')) 
-    {
-        BOSS_NAME = BOSS_NAME.replace(/\$/g, ' ');
-    }
-    BOSS_HEALTH = BOSS_HEALTH + Number(arr[4]);
-    HP_BAR_MAX_FRAME = Number(arr[5]);
-    HP_BAR_FRAME = Number(arr[5]);
-    if(Number(arr[5]) === 1)
-    {
-        BOSS_HUD_IND = false;
-    }
-    BOSS_PERCENT_C = arr[6];
-    if(BOSS_PERCENT_C == null)
-    {
-        BOSS_PERCENT_C = "";
-    }
-    IS_BOSS_FIGHT = true;
-    Instance.EntFireAtName({ name: BOSS_SCRIPT, input: "RunScriptInput", value: "CheckHealth", delay: 0.00 });
-}
-
-function AddHealth(arg)
-{
-    let arg_s = arg;
-    let arg_rs = arg_s.replace(/\s+/g, '');
-    const arr = arg_rs.split(",");
-    if(arr[0] == "0")
-    {
-        let players = Instance.FindEntitiesByClass("player");
-        if(players.length > 0)
-        {
-            for (let i = 0; i < players.length; i++) 
-            {
-                if(IsValidEntity(players[i]))
-                {
-                    BOSS_HEALTH = BOSS_HEALTH + Number(arr[1]);
-                }
-            }
-        }
-    }
-    else
-    {
-        BOSS_HEALTH = BOSS_HEALTH + Number(arr[1]);
-    }
-}
-
-Instance.OnScriptInput("CheckHealth", () => {
-    if(!IS_BOSS_FIGHT)
-    {
-        return;
-    }
-
-    if(HP_PER_FRAME == 0)
-    {
-        HP_PER_FRAME = BOSS_HEALTH / HP_BAR_MAX_FRAME;
-        BOSS_MAX_HEALTH = BOSS_HEALTH;
-    }
-
-    if(BOSS_HEALTH > BOSS_MAX_HEALTH)
-    {
-        HP_PER_FRAME = BOSS_HEALTH / HP_BAR_MAX_FRAME;
-        BOSS_MAX_HEALTH = BOSS_HEALTH;
-    }
-
-    HP_BAR_FRAME = BOSS_HEALTH / HP_PER_FRAME;
-    if(HP_BAR_FRAME > HP_BAR_MAX_FRAME)
-    {
-        HP_BAR_FRAME = HP_BAR_MAX_FRAME;
-    }
-
-    if(BOSS_HEALTH <= 0)
-    {
-        BOSS_HEALTH = 0;
-        Instance.EntFireAtName({ name: BOSS_SCRIPT, input: "RunScriptInput", value: "BossKill", delay: 0.00 });
-        return;
-    }
-    BuildHud();
-    Instance.EntFireAtName({ name: BOSS_SCRIPT, input: "RunScriptInput", value: "CheckHealth", delay: TICKRATE_B });
-});
-
-function BuildHud()
-{
-    if(!IS_BOSS_FIGHT)
-    {
-        return;
-    }
-    BOSS_HUD_TEXT = "";
-    let GrenadeDamage_String = "";
-    if(ITEM_DAMAGE != "")
-    {
-        ITEM_DAMAGE_TICK = ITEM_DAMAGE_TICK - TICKRATE_B;
-    }
-    if(ITEM_DAMAGE_TICK <= 0)
-    {
-        ITEM_DAMAGE = "";
-        ITEM_DAMAGE_TICK = SAVE_ITEM_DAMAG_T;
-    }
-
-    if(GRENADE_DAMAGE != 0)
-    {
-        GrenadeDamage_String = " [HE: -" + GRENADE_DAMAGE + " HP] ";
-        GRENADE_DAMAGE_TICK = GRENADE_DAMAGE_TICK - TICKRATE_B;
-    }
-    if(GRENADE_DAMAGE_TICK <= 0)
-    {
-        GRENADE_DAMAGE = 0;
-        GRENADE_DAMAGE_TICK = SAVE_GRENADE_DAMAG_T;
-    }
-    if(BOSS_HEALTH < 0)
-    {
-        BOSS_HEALTH = 0;
-    }
-    let PERCENT_HP = Math.ceil(BOSS_HEALTH / BOSS_MAX_HEALTH * 100);
-    BOSS_HUD_TEXT += `${BOSS_NAME}: ${BOSS_HEALTH} (${PERCENT_HP}%)${GrenadeDamage_String}${ITEM_DAMAGE}`;
-    if(BOSS_PERCENT_C.length > 0)
-    {
-        Instance.EntFireAtName({ name: BOSS_PERCENT_C, input: "InValue", value: ""+PERCENT_HP, delay: 0.00 });
-    }
-    if(BOSS_HUD_IND)
-    {
-        BOSS_HUD_TEXT += "\n[";
-        let hp_bar_int = Math.ceil(HP_BAR_FRAME);
-        if(hp_bar_int < 1)
-        {
-            hp_bar_int = 1;
-        }
-        for(let c = 0; c < hp_bar_int; c++)
-        {
-            BOSS_HUD_TEXT = BOSS_HUD_TEXT + BOSS_HUD_ST;
-        }
-        if(hp_bar_int < HP_BAR_MAX_FRAME)
-        {
-            for(let a = hp_bar_int; a < HP_BAR_MAX_FRAME; a++)
-            {
-                BOSS_HUD_TEXT = BOSS_HUD_TEXT + BOSS_HUD_ST2;
-            }
-        }
-        BOSS_HUD_TEXT = BOSS_HUD_TEXT + "]";
-    }
-    Instance.EntFireAtName({ name: BOSS_HUD_ENT, input: "SetMessage", value: BOSS_HUD_TEXT, delay: 0.00 });
-    //Instance.Msg(BOSS_HUD_TEXT);
-}
-
-Instance.OnScriptInput("SubtractHealth", () => {
-    if(!IS_BOSS_FIGHT)
-    {
-        return;
-    }
-        
-    if(BOSS_HEALTH >= 0)
-    {
-        BOSS_HEALTH = BOSS_HEALTH - 1
-    }
-});
-
-function ChangeHealthIt(arg)
-{
-    if(!IS_BOSS_FIGHT )
-    {
-        return;
-    }
-    if(BOSS_HEALTH >= 0)
-    {
-        BOSS_HEALTH = BOSS_HEALTH - arg;
-    }
-}
-
-function GrenadeDamage(arg)
-{
-    if(!IS_BOSS_FIGHT )
-    {
-        return;
-    }
-    if(BOSS_HEALTH >= 0)
-    {
-        BOSS_HEALTH = BOSS_HEALTH - arg;
-    }
-   GRENADE_DAMAGE = Number(GRENADE_DAMAGE) + Number(arg);
-   GRENADE_DAMAGE_TICK = 2.00;
-}
-
-function ItemDamage(arg)
-{
-    if(!IS_BOSS_FIGHT )
-    {
-        return;
-    }
-    let arg_s = arg;
-    let arg_rs = arg_s.replace(/\s+/g, '');
-    const arr = arg_rs.split(",");
-    let damage = Number(arr[1]);
-    let subs = "-";
-    if(BOSS_HEALTH >= 0)
-    {
-        BOSS_HEALTH = BOSS_HEALTH - damage;
-    }
-    if(damage < 0)
-    {
-        subs = "+";
-    }
-    ITEM_DAMAGE = " (" + arr[0] + ": "+subs+""+ Math.abs(damage) + " HP) ";
-}
-
-
-Instance.OnScriptInput("BossKill", () => {
-    BOSS_HEALTH = 0.00;
-    IS_BOSS_FIGHT = false;
-    Instance.EntFireAtName({ name: BOSS_ENT, input: "FireUser3", value: "", delay: 0.00 });
-    Instance.EntFireAtName({ name: BOSS_HUD_ENT, input: "SetMessage", value: BOSS_NAME+": 0", delay: 0.00 });
-    Instance.EntFireAtName({ name: BOSS_HUD_ENT, input: "HideHudHint", value: "", delay: 0.02 });
-    ResetBossS();
-});
-
-function IsValidEntity(ent)
-{
-    if(ent?.IsValid() && ent?.GetHealth() > 0 && ent?.GetTeamNumber() == 3)
-    {
-        return true;
-    }
-    return false;
-}
-
 //    ___ _       _           _     ___                 _   _                 
 //   / _ \ | ___ | |__   __ _| |   / __\   _ _ __   ___| |_(_) ___  _ __  ___ 
 //  / /_\/ |/ _ \| '_ \ / _` | |  / _\| | | | '_ \ / __| __| |/ _ \| '_ \/ __|
@@ -3234,25 +2967,25 @@ function GetValidPlayersCT()
 
 function NAVMESH_GetNearestNavPoint(vecOrigin)
 {
-    let ID = -1;
-    let iMin = 99999;
-    for (let i = 0; i < NAV_POINT_LIST.length; i++)
-    {
-        if (Vector3Utils.distance(vecOrigin, NAV_POINT_LIST[i].origin) > 128)
-        {
-            continue;
-        }
+	let ID = -1;
+	let iMin = 99999;
+	for (let i = 0; i < NAV_POINT_LIST.length; i++)
+	{
+		if (Vector3Utils.distance(vecOrigin, NAV_POINT_LIST[i].origin) > 128)
+		{
+			continue;
+		}
 
-        const iDistance = Vector3Utils.distance(vecOrigin, NAV_POINT_LIST[i].origin);
+		const iDistance = Vector3Utils.distance(vecOrigin, NAV_POINT_LIST[i].origin);
 
-        if (iDistance < iMin)
-        {
-            iMin = iDistance;
-            ID = i;
-        }
-    }
+		if (iDistance < iMin)
+		{
+			iMin = iDistance;
+			ID = i;
+		}
+	}
 
-    return ID;
+	return ID;
 }
 
 function Delay(callback, delaySeconds) {
@@ -3327,7 +3060,7 @@ function ResetVariables()
     if(isExtremeMode)
     {
         // VALUES
-        pre_human_hp = 120;
+        pre_human_hp = 130;
         pre_human_max_hp = 145;
         pre_traps_percentage = 80;
         pre_npcs_percentage = 50;
@@ -3373,24 +3106,6 @@ function ResetMusicList()
         let music = MUSIC_LIST_MAIN[i]
         MUSIC_LIST.push(music)
     }
-}
-
-function ResetBossS()
-{
-    BOSS_HEALTH = 0.00;
-    BOSS_MAX_HEALTH = 0.00;
-    HP_BAR_MAX_FRAME = 15;
-    HP_BAR_FRAME = 0;
-    HP_PER_FRAME = 0;
-    BOSS_NAME = "BOSS: ";
-    BOSS_ENT = "";
-    BOSS_SCRIPT = "";
-    BOSS_HUD_ENT = "";
-    BOSS_HUD_TEXT = "";
-    IS_BOSS_FIGHT = false;
-    BOSS_HUD_IND = true;
-    ITEM_DAMAGE = "";
-    GRENADE_DAMAGE = 0;
 }
 
 function ResetScript()
